[#64,view/map] Add small editor for user input
Right now the form lets you define the fields where the geometries are
stored (either in a single GeoJSON field or a pair of lat/lon fields).

When adding features, if wrong geometries are found (ie wrong field
values), the process will stop and an error notice will be shown.