[dataproxy][s] Rename duplicate fieldIds from dataproxy results.
Allow the dataproxy backend to be used for viewing files where
the column names are not unique.  This occurs, for example, when
more than 1 column has not been given a column title at all.

Prior to this commit, when encountering such a file, the dataproxy
backend would error-out leaving the user without any feedback as
to what went wrong.