Allow children in mobile nav
Adds some finishing touches to @olayway's work to
allow the mobile nav to render arbitrary children
which can be useful for things like site-wide
navigation.

Also adds some stories to demonstrate the new
functionality and migrates to the Storybook NextJs
'framework' which allows the Nav component to be
rendered in a story with all the routing
automatically stubbed and taken care of.